wip: refactor tests
