+++ src/lib/__tests__/stack.test.ts
@@ -1,6 +1,7 @@
 import { faker } from '@faker-js/faker/locale/af_ZA'
 import { Stack } from '@src/lib/stack'
 import { TerraformStack, App as TerraformApp } from 'cdktf'
+import { isString } from 'lodash'
 
 describe('Stack', () => {
   let stackName: string
@@ -37,7 +38,17 @@ describe('Stack', () => {
 
   it('exports the stack as a terraform object', () => {
     const output = subject.toSynthesized()
+
     expect(output).toMatchObject({
+      name: stackName,
+      dependencies: [],
+      constructPath: expect.stringMatching(`/${stackName}$`),
+      synthesizedStackPath: expect.stringMatching(`/${stackName}/main.tf.json$`),
+      workingDirectory: expect.stringMatching(`/${stackName}$`),
+    })
+
+    expect(isString(output.content)).toBe(true)
+    expect(JSON.parse(output.content)).toMatchObject({
       '//': {
         metadata: { stackName },
         outputs: {},
