+++ src/services/providers/aws/services/__tests__/networking.test.ts
@@ -1,7 +1,7 @@
 import { internetGateway, subnet, vpc } from '@cdktf/provider-aws'
 import { REGIONS } from '@aws/constants'
 import { DEFAULT_PROFILE_NAME, PROVIDER, SERVICE_TYPE } from '@src/constants'
-import { getProvisionable } from '@tests/mocks/getProvisionable'
+import { getProvisionable } from '@tests/helpers/getProvisionable'
 import { Stack } from '@lib/stack'
 import { faker } from '@faker-js/faker'
 import { TerraformOutput } from 'cdktf'
