+++ src/services/providers/local/services/__tests__/provider.test.ts
@@ -2,7 +2,7 @@ import { provider as terraformLocalProvider } from '@cdktf/provider-local'
 import { PROVIDER, SERVICE_TYPE } from '@src/constants'
 import { Stack } from '@lib/stack'
 import { LocalProvider } from '@src/services/providers/local/services/provider'
-import { getProvisionable } from '@tests/mocks/getProvisionable'
+import { getProvisionable } from '@tests/helpers/getProvisionable'
 import type { LocalProviderAttributes } from '@local/types'
 
 describe('Local Provider', () => {
