+++ src/services/providers/local/services/__tests__/state.test.ts
@@ -2,7 +2,7 @@ import { LocalBackend } from 'cdktf'
 import { LocalState } from '@src/services/providers/local/services/state'
 import { PROVIDER, SERVICE_TYPE } from '@src/constants'
 import { Stack } from '@lib/stack'
-import { getProvisionable } from '@tests/mocks/getProvisionable'
+import { getProvisionable } from '@tests/helpers/getProvisionable'
 import type { LocalStateAttributes } from '@local/services/state'
 
 describe('Local state', () => {
