+++ tests/helpers/getAwsServicePrerequisites.ts
