--- tests/mocks/index.ts
@@ -1,3 +1,0 @@
-export * from './getAwsCacheMock'
-export * from './getAwsDbMock'
-export * from './getAwsProvisionable'
